Group AVSC registration support by format
Move AVSC JSON parsing and schema registration helpers into the Avsc area, leaving Registry focused on registry state and options. Remove the old source/protocol extension split and keep default value helpers in shared extensions.
